requirements.txt
hydroloader >= 0.1.9
APScheduler == 3.10.1
pystray >= 0.19.4
pytz >= 2023.3
Pillow == 9.5.0
appdirs >= 1.4.4
customtkinter == 5.2.0
requests >= 2.27.1
tkinter-tooltip == 2.1.0